Step 1 of dynamic go bot refactor.
