+++ go.mod
@@ -3,24 +3,16 @@ module github.com/mattermost/mattermost-bot-sample-golang
 go 1.15
 
 require (
-	github.com/NYTimes/gziphandler v1.1.1 // indirect
-	github.com/go-gorp/gorp v2.0.0+incompatible // indirect
 	github.com/google/uuid v1.3.0 // indirect
 	github.com/gorilla/websocket v1.5.0 // indirect
 	github.com/klauspost/compress v1.15.2 // indirect
 	github.com/klauspost/cpuid/v2 v2.0.12 // indirect
-	github.com/mattermost/gorp v2.0.1-0.20190301154413-3b31e9a39d05+incompatible // indirect
 	github.com/mattermost/ldap v3.0.4+incompatible // indirect
 	github.com/mattermost/mattermost-server/v5 v5.39.3
-	github.com/mattermost/viper v1.0.4 // indirect
-	github.com/minio/minio-go/v6 v6.0.40 // indirect
 	github.com/minio/minio-go/v7 v7.0.24 // indirect
 	github.com/pelletier/go-toml v1.9.5 // indirect
 	github.com/rs/xid v1.4.0 // indirect
-	github.com/segmentio/analytics-go v3.1.0+incompatible // indirect
-	github.com/spf13/afero v1.2.2 // indirect
-	github.com/spf13/jwalterweatherman v1.1.0 // indirect
-	github.com/tkanos/gonfig v0.0.0-20210106201359-53e13348de2f // indirect
+	github.com/tkanos/gonfig v0.0.0-20210106201359-53e13348de2f
 	go.uber.org/atomic v1.9.0 // indirect
 	go.uber.org/multierr v1.8.0 // indirect
 	go.uber.org/zap v1.21.0 // indirect
@@ -29,5 +21,4 @@ require (
 	golang.org/x/sys v0.0.0-20220503163025-988cb79eb6c6 // indirect
 	gopkg.in/asn1-ber.v1 v1.0.0-20181015200546-f715ec2f112d // indirect
 	gopkg.in/ini.v1 v1.66.4 // indirect
-	gopkg.in/olivere/elastic.v5 v5.0.82 // indirect
 )
